Finished Back-end Inventory
There are 3 layer of implementing the Inventory System.

1st layer(outer-most layer): Inventory Manager
Holder of the constructor, can be served as different storage stations such as a storage chest, loot-box or player inventory, etc.

2nd : Inventory Constructor
Holding a List of inventory slot, controls the logic of distributing collected item into desired slot.

3rd: Inventory Slot
controls the logic of a single slot, add or remove stack of item.